Got a real-world JVM cpu-sampling profile to serialize and come over to the mock-backend over the wire. The test does not pass yet, it does get an aggregated data-structure now, but the stack is backwards etc.

+++ e2etest/src/test/java/fk/prof/recorder/WorkHandlingTest.java
@@ -351,7 +351,7 @@ public class WorkHandlingTest {
             csum.reset();
             csum.update(req, bytesAfterChksum, bytesBeforeChksum);
             int wseChksum = is.readUInt32();
-            assertThat(wseChksum, is(csum.getValue()));
+            assertThat(wseChksum, is((int) csum.getValue()));
             bytesAfterChksum = is.getTotalBytesRead();
             ///////////////////////
         }
